+++ pretext/utils.py
@@ -259,6 +259,20 @@ def serve_forever(
             log.warning(f"Trying port {port} instead.\n")
 
 
+def start_codespace_server(
+    access: t.Literal["public", "private"] = "private",
+    port: int = 8128,
+) -> None:
+    """
+    Temporary hack until we can figure out what is going on with codespaces an the (possibly more robust) server process we usually define.
+    """
+    subprocess.run(
+        f"python -m http.server {port}",
+        shell=True,
+    )
+    return
+
+
 # Info on namespaces: http://lxml.de/tutorial.html#namespaces
 NSMAP = {
     "xi": "http://www.w3.org/2001/XInclude",